fix rest of share functionality, change setting button permissions
