+++ priv/gettext/default.pot
@@ -41,7 +41,7 @@ msgstr ""
 msgid "Are you sure you want to delete this idea?"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:188
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:173
 #, elixir-autogen, elixir-format
 msgid "Attention: This will delete the brainstorming with all belonging ideas and other associated records to it. This cant be undone"
 msgstr ""
@@ -72,22 +72,22 @@ msgstr ""
 msgid "Create!"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:198
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:183
 #, elixir-autogen, elixir-format
 msgid "Delete"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:182
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:167
 #, elixir-autogen, elixir-format
 msgid "Delete Brainstorming"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:146
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:131
 #, elixir-autogen, elixir-format
 msgid "Export to CSV"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:150
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:135
 #, elixir-autogen, elixir-format
 msgid "Export to HTML"
 msgstr ""
@@ -178,7 +178,7 @@ msgstr ""
 msgid "Edit Brainstorming"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:139
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:124
 #, elixir-autogen, elixir-format
 msgid "Export"
 msgstr ""
@@ -211,42 +211,32 @@ msgstr ""
 msgid "Administration for brainstorming: %{name}"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:67
-#, elixir-autogen, elixir-format
-msgid "Show brainstorming settings link for all users"
-msgstr ""
-
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:71
-#, elixir-autogen, elixir-format
-msgid "Warning: Please make sure you save the admin link at the top, before hiding the settings link!"
-msgstr ""
-
 #: lib/mindwendel_web/controllers/static_page_html/home.html.heex:37
 #, elixir-autogen, elixir-format
 msgid "Brainstormings will be deleted after %{days} days."
 msgstr ""
 
-#: lib/mindwendel/brainstormings/brainstorming.ex:75
+#: lib/mindwendel/brainstormings/brainstorming.ex:73
 #, elixir-autogen, elixir-format
 msgid "cyan"
 msgstr ""
 
-#: lib/mindwendel/brainstormings/brainstorming.ex:76
+#: lib/mindwendel/brainstormings/brainstorming.ex:74
 #, elixir-autogen, elixir-format
 msgid "gray dark"
 msgstr ""
 
-#: lib/mindwendel/brainstormings/brainstorming.ex:77
+#: lib/mindwendel/brainstormings/brainstorming.ex:75
 #, elixir-autogen, elixir-format
 msgid "green"
 msgstr ""
 
-#: lib/mindwendel/brainstormings/brainstorming.ex:78
+#: lib/mindwendel/brainstormings/brainstorming.ex:76
 #, elixir-autogen, elixir-format
 msgid "red"
 msgstr ""
 
-#: lib/mindwendel/brainstormings/brainstorming.ex:79
+#: lib/mindwendel/brainstormings/brainstorming.ex:77
 #, elixir-autogen, elixir-format
 msgid "yellow"
 msgstr ""
@@ -280,17 +270,17 @@ msgstr ""
 msgid "Share brainstorming"
 msgstr ""
 
-#: lib/mindwendel_web/live/brainstorming_live/share_component.html.heex:80
+#: lib/mindwendel_web/live/brainstorming_live/share_component.html.heex:82
 #, elixir-autogen, elixir-format
 msgid "Download as png"
 msgstr ""
 
-#: lib/mindwendel_web/live/brainstorming_live/share_component.html.heex:69
+#: lib/mindwendel_web/live/brainstorming_live/share_component.html.heex:71
 #, elixir-autogen, elixir-format
 msgid "Download as svg"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:194
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:179
 #, elixir-autogen, elixir-format
 msgid "Brainstorming delete are you sure"
 msgstr ""
@@ -300,18 +290,18 @@ msgstr ""
 msgid "Update idea"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:163
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:148
 #, elixir-autogen, elixir-format
 msgid "Attention: This will delete all ideas in this brainstorming. The brainstorming itself and labels will be left unchanged. This cant be undone."
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:171
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:156
 #, elixir-autogen, elixir-format
 msgid "Brainstorming will be emptied - are you sure"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:157
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:173
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:142
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:158
 #, elixir-autogen, elixir-format
 msgid "Empty brainstorming"
 msgstr ""
@@ -326,7 +316,7 @@ msgstr ""
 msgid "Attention: Brainstormings will be deleted %{available_until} after last access!"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:81
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:66
 #, elixir-autogen, elixir-format
 msgid "Allow users to change the order of ideas"
 msgstr ""
@@ -377,27 +367,27 @@ msgstr ""
 msgid "close"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:129
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:114
 #, elixir-autogen, elixir-format
 msgid "Add idea label"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:101
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:86
 #, elixir-autogen, elixir-format
 msgid "Choose the label color"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:90
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:75
 #, elixir-autogen, elixir-format
 msgid "Edit Brainstorming Labels"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:117
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:102
 #, elixir-autogen, elixir-format
 msgid "Remove idea label"
 msgstr ""
 
-#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:109
+#: lib/mindwendel_web/live/admin/brainstorming_live/edit.html.heex:94
 #, elixir-autogen, elixir-format
 msgid "Type the label name"
 msgstr ""
@@ -572,4 +562,9 @@ msgstr ""
 #: lib/mindwendel_web/live/brainstorming_live/show.html.heex:97
 #, elixir-autogen, elixir-format
 msgid "Idea details"
+msgstr ""
+
+#: lib/mindwendel_web/live/brainstorming_live/share_component.html.heex:45
+#, elixir-autogen, elixir-format
+msgid "Änderungen erlauben"
 msgstr ""
